Added new webservice to get the drops

+++ LocalChat/src/main/javas/com/chat/models/Location.java
@@ -2,29 +2,29 @@ package com.chat.models;
 
 public class Location {
 	
-	private Float lat;
-	private Float lon;
+	private Double lat;
+	private Double lon;
 	
 	public Location() {
 		
 	}
 	
-	public Location(float lat, float lon) {
+	public Location(Double lat, Double lon) {
 		
 		this.lat = lat;
 		this.lon = lon;
 	}
 	
-	public float getLat() {
+	public Double getLat() {
 		return lat;
 	}
-	public void setLat(float lat) {
+	public void setLat(Double lat) {
 		this.lat = lat;
 	}
-	public float getLon() {
+	public Double getLon() {
 		return lon;
 	}
-	public void setLon(float lon) {
+	public void setLon(Double lon) {
 		this.lon = lon;
 	}
